BUILD: `base` no longer depends on `parser`, move classes
