+++ base/build.gradle.kts
@@ -4,7 +4,6 @@ dependencies {
   api(project(":api"))
   api(project(":tools"))
   api(project(":pretty"))
-  api(project(":parser"))
   val deps: java.util.Properties by rootProject.ext
   implementation("org.aya-prover", "commonmark", version = deps.getProperty("version.commonmark"))
   testImplementation("org.junit.jupiter", "junit-jupiter", version = deps.getProperty("version.junit"))
