+++ base/src/main/java/module-info.java
@@ -1,19 +1,16 @@
 module org.aya {
   requires static org.jetbrains.annotations;
-  requires static org.antlr.antlr4.runtime;
+
   requires org.commonmark;
 
   requires transitive org.aya.api;
-  requires transitive org.aya.parser;
   requires transitive org.aya.pretty;
   requires transitive org.aya.util;
   requires transitive kala.base;
   requires transitive kala.collection;
 
-  exports org.aya.concrete.desugar.error;
   exports org.aya.concrete.desugar;
-  exports org.aya.concrete.parse.error;
-  exports org.aya.concrete.parse;
+  exports org.aya.concrete.error;
   exports org.aya.concrete.remark;
   exports org.aya.concrete.resolve.context;
   exports org.aya.concrete.resolve.error;
