+++ base/src/main/java/org/aya/concrete/desugar/AyaBinOpSet.java
@@ -6,7 +6,7 @@ import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.MutableMap;
 import org.aya.api.error.Reporter;
 import org.aya.api.ref.DefVar;
-import org.aya.concrete.desugar.error.OperatorProblem;
+import org.aya.concrete.error.OperatorProblem;
 import org.aya.concrete.resolve.context.Context;
 import org.aya.util.binop.BinOpSet;
 import org.aya.util.binop.OpDecl;
