+++ base/src/main/java/org/aya/concrete/desugar/BinExprParser.java
@@ -5,7 +5,7 @@ package org.aya.concrete.desugar;
 import kala.collection.SeqView;
 import org.aya.api.ref.DefVar;
 import org.aya.concrete.Expr;
-import org.aya.concrete.desugar.error.OperatorProblem;
+import org.aya.concrete.error.OperatorProblem;
 import org.aya.generic.Constants;
 import org.aya.pretty.doc.Doc;
 import org.aya.util.binop.Assoc;
