+++ base/src/main/java/org/aya/concrete/resolve/visitor/StmtResolver.java
@@ -10,7 +10,7 @@ import org.aya.api.ref.DefVar;
 import org.aya.api.ref.LocalVar;
 import org.aya.concrete.Pattern;
 import org.aya.concrete.desugar.AyaBinOpSet;
-import org.aya.concrete.desugar.error.OperatorProblem;
+import org.aya.concrete.error.OperatorProblem;
 import org.aya.concrete.remark.Remark;
 import org.aya.concrete.resolve.ResolveInfo;
 import org.aya.concrete.resolve.context.Context;
