+++ cli/build.gradle.kts
@@ -6,6 +6,7 @@ CommonTasks.fatJar(project, "org.aya.cli.Main")
 
 dependencies {
   api(project(":base"))
+  api(project(":parser"))
   api(project(":tools-repl"))
   val deps: java.util.Properties by rootProject.ext
   api("com.google.code.gson", "gson", version = deps.getProperty("version.gson"))
